requirements.txt
customtkinter~=5.2.2

openpyxl~=3.1.5
pypdf~=5.0.1
pillow~=10.4.0